backend/requirement.txt
fastapi
uvicorn
requests
python-dotenv
PyGithub
langgraph
langchain
langchain-groq  
pydantic
redis
